adaugat descrieri functii. updatat ins si del pentru cazuri speciale + modificat returnare. creat chg. adaugat micro.h pt microinstructiuni.
